+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to `homebridge-config-ui-x` will be documented in this file.
 ### UI Changes
 
 - upgrade `xterm` to `v6` + style improvements
+- added option: global terminal/log window settings
 
 ### Other Changes